refactor(transpiler): extract parse-unparse.mts that is more generic
